# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Minilog API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/household": {
      "get": {
        "tags": [
          "household"
        ],
        "summary": "Get Household",
        "operationId": "get_household_api_v1_household_get",
        "parameters": [
          {
            "name": "minilog_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minilog Session"
            }
          },
          {
            "name": "minilog_csrf",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minilog Csrf"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "household"
        ],
        "summary": "Delete Household",
        "operationId": "delete_household_api_v1_household_delete",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "minilog_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minilog Session"
            }
          },
          {
            "name": "minilog_csrf",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minilog Csrf"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HouseholdDeleteRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/exports/minilog": {
      "get": {
        "tags": [
          "exports"
        ],
        "summary": "Export Minilog",
        "operationId": "export_minilog_api_v1_exports_minilog_get",
        "parameters": [
          {
            "name": "minilog_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minilog Session"
            }
          },
          {
            "name": "minilog_csrf",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minilog Csrf"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exports/timeline.csv": {
      "get": {
        "tags": [
          "exports"
        ],
        "summary": "Export Timeline Csv",
        "operationId": "export_timeline_csv_api_v1_exports_timeline_csv_get",
        "parameters": [
          {
            "name": "baby_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Baby Id"
            }
          },
          {
            "name": "minilog_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minilog Session"
            }
          },
          {
            "name": "minilog_csrf",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minilog Csrf"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClockFormat": {
        "type": "string",
        "enum": [
          "12h",
          "24h"
        ],
        "title": "ClockFormat"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HouseholdDeleteRequest": {
        "properties": {
          "confirmation": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Confirmation"
          }
        },
        "type": "object",
        "required": [
          "confirmation"
        ],
        "title": "HouseholdDeleteRequest"
      },
      "HouseholdOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "time_zone": {
            "type": "string",
            "title": "Time Zone"
          },
          "locale": {
            "type": "string",
            "title": "Locale"
          },
          "clock_format": {
            "$ref": "#/components/schemas/ClockFormat"
          },
          "measurement_system": {
            "$ref": "#/components/schemas/MeasurementSystem"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_name",
          "time_zone",
          "locale",
          "clock_format",
          "measurement_system"
        ],
        "title": "HouseholdOut"
      },
...
      "MeasurementSystem": {
        "type": "string",
        "enum": [
          "metric",
          "imperial"
        ],
        "title": "MeasurementSystem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}